Resolve flipping-between-frames issues
Solves issues described in 93fe0b with two changes:
 1. When queueing rebuild of field, don't queue redraw at same time, but wait
    till after rebuilt. Solves that when hovering over button/slider, window
    flipped to previous frame.
 2. Draw frames twice consecutively. Solves that when maximizied, last
    frames still shown (with prior appearance of not-maximized window
    visible) (Ubuntu), and solves Windows issue of last two frames being
    flipped between.

Elaboration on 2.:
A similar problem described in 93fe0b happened on Window 10 (as opposed
to Ubuntu 14.04LTS): the screen would flip between the last two
frames rapidly, thoughout execution. E.g., when window maximized, (this part similar
to Ubuntu issue), the last frame had the window smaller, so it flipped between the
current frame and a frame with the unmaximized window in view. This would only
include the conrod GUI when .draw_if_changed was also only called upon
changes to the field view, but not normally. Again, this is fixed
by this commit, albeit suboptimally. There should be a better solution.