Fix color-shading
This removes an unnecessary case in the color creation for if
potential is below zero, which it will never, significantly be.

Additionally, some arrows in the two-negative-charge-field have potential
values just below zero (e.g., -0.002...) due to float imprecision, resulting
in a negative value in the R channel, which made some arrows that should have
been deep blue, very red instead. (Fixed)